[Unit tests] Ignoring empty tournaments in tournament_injector teardown

+++ daoserver/src/unit_tests/tournament_injector.py
@@ -96,6 +96,9 @@ class TournamentInjector(object):
         for tourn in Tournament.query.filter(Tournament.id.in_(
             tuple(self.tournament_ids))).all():
 
+            if tourn is None:
+                return
+
             for rnd in tourn.rounds:
                 rnd.round_scores.delete()
